Architecture update
Added the component lib to hold packages of the scientific method library sciMeth. Also started to add contents to the SciMethOperation package. Small changes to the Document class of the SciMethData package that now can hold a collection of ScientificAnalysis. Pending to define how the sequence of operations is stored and executed.