Update installer v0.1
We continue using pynsist. I corrected several issues with installer. Previously, although the installer appears to work, when running the program afterwards, the application would complain of not finding the libraries, and when further added hidden imports (PyQt5.sip). All libraries are now correctly found and the entry point is findable whether from scripting mode (i.e. from your IDE of taste) or from an installed version.
Still, the direct link to the launcher won't run adequately failing to link to python.exe. Instead I have prepared a bypass with a .bat file, but even this will require the use to make a direct link to the .bat file. With regards to this latter, pynsist indicates that the working directory of shortcut cannot be directly manipulated but manual editing of the nsist .nsi file is require. Hence, by now, the user will have to create this direct link by hand.