Manual incorporation of tkinter
The tkinter GUI package is necessary for matplotlib. It was giving us a lot of trouble for packaging the installer with pynsist since is unable to get it automatically. See:

https://pynsist.readthedocs.io/en/latest/faq.html#packaging-with-tkinter

WATCH OUT! Since the tkinter version "copied" under the above instructions has to be the binaries corresponding to the version of python being used, if I ever upgrade my python, I'll have to also manually "recopy" the tkinter library files according to those instructions.